Task/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
